requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.0.0=pypi_0
albumentations=1.3.1=pypi_0
av=10.0.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=5.3.1=pypi_0
certifi=2022.12.7=py38h06a4308_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
configparser=5.3.0=pypi_0
cryptography=38.0.4=py38h9ce1e76_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.11.0=pypi_0
cython=0.29.33=pypi_0
ffmpeg=4.3=hf484d3e_0
flit-core=3.6.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
gdown=4.7.1=pypi_0
giflib=5.2.1=h5eee18b_1
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.23.3=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.59.0=pypi_0
h5py=2.10.0=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.25.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
joblib=1.2.0=pypi_0
jpeg=9e=h7f8727e_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.5.0=h6a678d5_1
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
lz4-c=1.9.4=h6a678d5_0
markdown=3.5=pypi_0
markupsafe=2.1.3=pypi_0
mat73=0.59=pypi_0
matplotlib=3.1.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.0=pypi_0
numpy=1.22.0=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.5.2.54=pypi_0
opencv-python-headless=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
pandas=1.1.5=pypi_0
pillow=9.3.0=py38h6a678d5_2
pip=22.3.1=py38h06a4308_0
protobuf=4.21.12=pypi_0
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pymysql=1.0.3=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.16=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
pytorch=1.12.1=py3.8_cuda10.2_cudnn7.6.5_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
qudida=0.0.4=pypi_0
readline=8.2=h5eee18b_0
requests=2.28.1=py38h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.17.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.5.2=pypi_0
setuptools=65.6.3=py38h06a4308_0
shutils=0.1.0=pypi_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.5=pypi_0
sqlite=3.40.1=h5082296_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.1=pypi_0
tensorboardx=2.4.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2023.2.3=pypi_0
tk=8.6.12=h1ccaba5_0
torchaudio=0.12.1=py38_cu102
torchvision=0.13.1=py38_cu102
tqdm=4.64.0=pypi_0
typing_extensions=4.4.0=py38h06a4308_0
urllib3=1.26.14=py38h06a4308_0
werkzeug=3.0.0=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h5eee18b_1
yacs=0.1.8=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
